Super basic extension
This is a baseline create react app that has been slightly modified to work as a chrome override - new tab.  I have wrapped the wrapStore in a timeout so it gets called after the new tab proxyStore.ready() is called.  This resulted in a null new tab on chrome start.  Every subsequent new tab is fine.